pyproject.toml
[project]
name = "formation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[all]>=0.115.12",
    "mlflow>=2.21.3",
    "pandas>=2.2.3",
    "pydantic-settings>=2.8.1",
    "ruff>=0.11.4",
    "scikit-learn>=1.6.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
